+++ app/src/main/java/page/ooooo/geoshare/MainActivity.kt
@@ -1,19 +1,11 @@
 package page.ooooo.geoshare
 
 import android.os.Bundle
-import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.platform.LocalContext
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dagger.hilt.android.AndroidEntryPoint
-import page.ooooo.geoshare.lib.ConversionFailed
-import page.ooooo.geoshare.lib.Message
-import page.ooooo.geoshare.lib.SharingSucceeded
 import page.ooooo.geoshare.ui.theme.AppTheme
 
 @AndroidEntryPoint
@@ -26,7 +18,7 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             AppTheme {
-                MainNavigation(viewModel)
+                MainNavigation(intent, viewModel)
             }
         }
     }
